pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memorix-sdk"
version = "0.1.0"
description = "A flexible memory management system for AI applications"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Memorix Team", email = "team@memorix.ai"}
]
maintainers = [
    {name = "Memorix Team", email = "team@memorix.ai"}
]
keywords = ["memory", "vector-store", "embeddings", "ai", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
openai = [
    "openai>=1.0.0",
]
gemini = [
    "google-generativeai>=0.3.0",
]
sentence-transformers = [
    "sentence-transformers>=2.0.0",
    "torch>=1.9.0",
]
faiss = [
    "faiss-cpu>=1.7.0",
]
qdrant = [
    "qdrant-client>=1.0.0",
]
all = [
    "memorix-sdk[openai,gemini,sentence-transformers,faiss,qdrant]",
]

[project.urls]
Homepage = "https://github.com/your-org/memorix-sdk"
Documentation = "https://docs.memorix.ai"
Repository = "https://github.com/your-org/memorix-sdk.git"
"Bug Tracker" = "https://github.com/your-org/memorix-sdk/issues"
"Discussions" = "https://github.com/your-org/memorix-sdk/discussions"

[project.scripts]
memorix = "memorix.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["memorix*"]

[tool.setuptools.package-data]
memorix = ["*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["memorix"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=memorix",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["memorix"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 